refactor: rename xmly_decryptor to xmly
